+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [TBA] - TBD
 ### Changed
 - Maven group ID is now `com.expediagroup.apiary` (was `com.expedia.apiary`).
+- All Java classes moved to `com.expediagroup.apiary` (was `com.expedia.apiary`).
 
 ## [2.0.0] - 2019-04-08
 ### Changed